pages/webtables.py
from pages.base_page import BasePage
from components.components import WebElement

class WebTables(BasePage):
    def __init__(self, driver):
        self.base_url = 'https://demoqa.com/webtables'
        self.no_rows = WebElement(driver, 'div.rt-noData')
        self.table_elements = WebElement(driver, '#app > div > div > div.row > div.col-12.mt-4.col-md-6 > div.web-tables-wrapper > div.ReactTable.-striped.-highlight > div.rt-table > div.rt-tbody > div:nth-child(1) > div > div:nth-child(1)')
        self.delete_button = WebElement(driver, 'span[title="Delete"]')
        super().__init__(driver, self.base_url)